Fix circular imports issue
