+++ src/pyscal/__version__.py
@@ -1,0 +1,6 @@
+try:
+    from .version import version
+
+    __version__ = version
+except ImportError:
+    __version__ = "0.0.0"
